Refactor floodFill tests and fix findClosestFood test case
