+++ __tests__/snake-logic.test.js
@@ -93,7 +93,7 @@ describe('snakeLogic', () => {
       };
       
       const isMoveSafe = { up: true, down: true, left: true, right: true };
-      expect(findClosestFood(game');
+      expect(findClosestFood(gameState, isMoveSafe)).toEqual('up');
     });
   });
